pulled module from testing

+++ api/src/main/java/com/community/api/endpoint/avisoft/controller/Category/CategoryController.java
@@ -64,8 +64,8 @@ public class CategoryController extends CatalogEndpoint {
                 return ResponseService.generateErrorResponse(CATALOGSERVICENOTINITIALIZED, HttpStatus.INTERNAL_SERVER_ERROR);
             }
 
-            if (addCategoryDto.getName() == null && addCategoryDto.getName().trim().isEmpty()) {
-                return new ResponseEntity<>("CATEGORY TITLE CANNOT BE EMPTY OR NULL", HttpStatus.BAD_REQUEST);
+            if (addCategoryDto.getName() == null || addCategoryDto.getName().trim().isEmpty()) {
+                return ResponseService.generateErrorResponse("CATEGORY TITLE CANNOT BE EMPTY OR NULL", HttpStatus.BAD_REQUEST);
             }
             addCategoryDto.setName(addCategoryDto.getName().trim());
             categoryImpl.setName(addCategoryDto.getName());
@@ -135,8 +135,8 @@ public class CategoryController extends CatalogEndpoint {
         }
     }
 
-    @GetMapping(value = "/get-products-by-category-id")
-    public ResponseEntity<?> getProductsByCategoryId(HttpServletRequest request, @RequestParam(value = "id") String id) throws Exception {
+    @GetMapping(value = "/get-products-by-category-id/{id}")
+    public ResponseEntity<?> getProductsByCategoryId(HttpServletRequest request, @PathVariable String id) {
         try {
             if (catalogService == null) {
                 return ResponseService.generateErrorResponse("CATALOG SERVICE IS NULL", HttpStatus.INTERNAL_SERVER_ERROR);
@@ -144,7 +144,7 @@ public class CategoryController extends CatalogEndpoint {
 
             Long categoryId = Long.parseLong(id);
             if (categoryId <= 0) {
-                return new ResponseEntity<>(CATEGORYCANNOTBELESSTHANOREQAULZERO, HttpStatus.INTERNAL_SERVER_ERROR);
+                return ResponseService.generateErrorResponse(CATEGORYCANNOTBELESSTHANOREQAULZERO, HttpStatus.INTERNAL_SERVER_ERROR);
             }
 
             Category category = this.catalogService.findCategoryById(categoryId);
@@ -173,7 +173,11 @@ public class CategoryController extends CatalogEndpoint {
 
             return ResponseService.generateSuccessResponse("CATEGORY DATA FOUND", categoryWrapper, HttpStatus.OK);
 
-        } catch (Exception exception) {
+        } catch (NumberFormatException numberFormatException) {
+            exceptionHandlingService.handleException(numberFormatException);
+            return ResponseService.generateErrorResponse(numberFormatException.getMessage(), HttpStatus.BAD_REQUEST);
+        }
+        catch (Exception exception) {
             exceptionHandlingService.handleException(exception);
             return ResponseService.generateErrorResponse(SOMEEXCEPTIONOCCURRED + ": " + exception.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -183,16 +187,16 @@ public class CategoryController extends CatalogEndpoint {
     public ResponseEntity<?> removeCategoryById(HttpServletRequest request, @PathVariable("categoryId") String id, @RequestParam(value = "productLimit", defaultValue = "20") int productLimit, @RequestParam(value = "productOffset", defaultValue = "1") int productOffset, @RequestParam(value = "subcategoryLimit", defaultValue = "20") int subcategoryLimit, @RequestParam(value = "subcategoryOffset", defaultValue = "1") int subcategoryOffset) {
         try {
             if (catalogService == null) {
-                return new ResponseEntity<>("CATALOG SERVICE IS NULL", HttpStatus.INTERNAL_SERVER_ERROR);
+                return ResponseService.generateErrorResponse("CATALOG SERVICE IS NULL", HttpStatus.INTERNAL_SERVER_ERROR);
             }
 
             Long categoryId = Long.parseLong(id);
             if (categoryId <= 0) {
-                return new ResponseEntity<>(CATEGORYCANNOTBELESSTHANOREQAULZERO, HttpStatus.INTERNAL_SERVER_ERROR);
+                return ResponseService.generateErrorResponse(CATEGORYCANNOTBELESSTHANOREQAULZERO, HttpStatus.INTERNAL_SERVER_ERROR);
             }
             Category category = this.catalogService.findCategoryById(categoryId);
 
-            if (category != null) {
+            if (category != null && ((Status)category).getArchived() != 'Y') {
 
                 catalogService.removeCategory(category);
                 return ResponseService.generateSuccessResponse("CATEGORY DELETED SUCCESSFULLY", "DELETED", HttpStatus.OK);
@@ -213,7 +217,7 @@ public class CategoryController extends CatalogEndpoint {
         try {
 
             if (catalogService == null) {
-                return new ResponseEntity<>(CATALOGSERVICENOTINITIALIZED, HttpStatus.INTERNAL_SERVER_ERROR);
+                return ResponseService.generateErrorResponse(CATALOGSERVICENOTINITIALIZED, HttpStatus.INTERNAL_SERVER_ERROR);
             }
             Long categoryId = Long.parseLong(id);
 
@@ -223,7 +227,7 @@ public class CategoryController extends CatalogEndpoint {
 
             Category category = this.catalogService.findCategoryById(categoryId);
 
-            if (category != null) {
+            if (category != null && ((Status)category).getArchived() != 'Y') {
 
                 if (!addCategoryDto.getName().isEmpty() && !addCategoryDto.getName().trim().isEmpty()) { // trim works on nonNull values only.
                     category.setName(addCategoryDto.getName().trim());
@@ -232,7 +236,7 @@ public class CategoryController extends CatalogEndpoint {
                     category.setDescription(addCategoryDto.getDescription().trim());
                 }
                 if (addCategoryDto.getActiveEndDate() != null && !addCategoryDto.getActiveEndDate().after(addCategoryDto.getActiveStartDate()) && !addCategoryDto.getActiveEndDate().after(new Date())) {
-                    return new ResponseEntity<>("ACTIVE END DATE CANNOT BE BEFORE OR EQUAL TO ACTIVE START DATE(CURRENT DATE)", HttpStatus.INTERNAL_SERVER_ERROR);
+                    return ResponseService.generateErrorResponse("ACTIVE END DATE CANNOT BE BEFORE OR EQUAL TO ACTIVE START DATE(CURRENT DATE)", HttpStatus.INTERNAL_SERVER_ERROR);
                 }
                 if (!addCategoryDto.getDisplayTemplate().isEmpty() && !addCategoryDto.getDisplayTemplate().trim().isEmpty()) {
                     category.setDisplayTemplate(addCategoryDto.getDescription().trim());
